chore: address review suggestions
